complete functionality of getting data by id

+++ SalarySystem-main/src/main/java/com/codeline/sampleProject/Models/Salary.java
@@ -28,7 +28,7 @@ public class Salary extends BaseEntity {
 
     Double healthCareContribution;
 
-    Double allowances;
+    Integer allowances;
 
     Double bonus;
